requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
bcrypt==4.1.2
PyJWT==2.8.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
email-validator==2.1.0